moving reusable scribbles code to ./monailabel/scribbles/.
