+++ monailabel/scribbles/__init__.py
@@ -8,14 +8,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from .activelearning import MyStrategy
-from .infer import SegmentationWithWriteLogits
-from .scribbles import (
-    SpleenInteractiveGraphCut,
-    SpleenISegCRF,
-    SpleenISegGraphCut,
-    SpleenISegGraphcutColdstart,
-    SpleenISegSimpleCRF,
-)
-from .train import MyTrain
