+++ monailabel/scribbles/utils.py
